add tenant migration primitives and model
