+++ vmngclient/primitives/tenant_management.py
@@ -1,53 +1,11 @@
-from datetime import datetime
-from typing import Dict, List, Optional, Union
+from typing import Dict, List
 
-from pydantic import BaseModel, Field, IPvAnyAddress
+from pydantic import BaseModel, Field
 
+from vmngclient.model.tenant import Tenant
 from vmngclient.primitives import APIPrimitiveBase, Versions, View
 from vmngclient.typed_list import DataSequence
 from vmngclient.utils.session_type import ProviderAsTenantView, ProviderView
-
-
-class MTEdge(BaseModel):
-    uuid: str
-    configured_hostname: Optional[str] = Field(alias="configuredHostname")
-    configured_system_ip: Optional[IPvAnyAddress] = Field(alias="configuredSystemIP")
-    management_system_ip: Optional[IPvAnyAddress] = Field(alias="managementSystemIP")
-    device_model: Optional[str] = Field(alias="deviceModel")
-    device_type: Optional[str] = Field(alias="deviceType")
-
-
-class TenantVPNMap(BaseModel):
-    tenant_vpn: int = Field(alias="tenantVPN")
-    device_vpn: int = Field(alias="deviceVPN")
-
-
-class Tenant(BaseModel):
-    name: str
-    org_name: str = Field(alias="orgName")
-    subdomain: str = Field(alias="subDomain")
-    flake_id: Optional[int] = Field(alias="flakeId")
-    vbond_address: Optional[str] = Field(alias="vBondAddress")
-    edge_connector_system_ip: Optional[str] = Field(alias="edgeConnectorSystemIp")
-    edge_connector_enable: Optional[bool] = Field(alias="edgeConnectorEnable")
-    vsmarts: List[str] = Field(alias="vSmarts", default=[])
-    wan_edge_forecast: Optional[int] = Field(alias="wanEdgeForecast")
-    saml_sp_info: Optional[str] = Field(alias="samlSpInfo")
-    idp_map: Union[Dict, str, None] = Field(alias="idpMap")
-    config_db_cluster_service_name: Optional[str] = Field(alias="configDBClusterServiceName")
-    old_idp_map: Union[Dict, str, None] = Field(alias="oldIdpMap")
-    created_at: Optional[datetime] = Field(alias="createdAt")
-    rid: Optional[int] = Field(alias="@rid")
-    edge_connector_tunnel_interface_name: Optional[str] = Field(alias="edgeConnectorTunnelInterfaceName")
-    tenant_id: Optional[str] = Field(alias="tenantId")
-    sp_metadata: Optional[str] = Field(alias="spMetadata")
-    state: Optional[str]
-    wan_edge_present: Optional[int] = Field(alias="wanEdgePresent")
-    desc: Optional[str]
-    mt_edge: Optional[List[MTEdge]] = Field(alias="mtEdge")
-    mt_edge_count: Optional[int] = Field(alias="mtEdgeCount")
-    tenant_vpn_map: Optional[List[TenantVPNMap]] = Field(alias="tenantVPNmap")
-    tenant_provider_vpn_count: Optional[int] = Field(alias="tenantProviderVPNCount")
 
 
 class TenantDeleteRequest(BaseModel):
